split the slow broker into the same three files as the non-slow broker
